Spawn camera when respawn
I tried your side scroller starter kit, and I've found a bug in multiplayer, whenever the player died, the camera returned to first person view. With this little piece of code, I could solve the problem. If you also think this is correct, I propose it to be added.

Another bug I noticed in multiplayer, for some reason player character stutters - only visually, however other players do not, I have no idea what causes it, maybe it's related to the movement prediction? Or to the camera?